clean up presenter to follow MVP better
